move BoardList.js to views
